+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ Change Log
 
 v4.2.1 (UNRELEASED)
 ------------------
-* :star: We added helpers to manage the alteration of the project_info attribute on the Scope object better.
+* :star: We added helpers and tests to manage the alteration of the project_info attribute on the Scope object better.
 
 v4.2.0 (10JUN22)
 ----------------